requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
azure-ai-translation-text==1.0.1
azure-core==1.32.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
cryptography==44.0.0
deepl==1.20.0
distro==1.9.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gradio==5.7.1
gradio_client==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
jiter==0.8.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.2
onnx==1.17.0
onnxruntime==1.20.1
openai==1.55.3
opencv-python-headless==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pdf2zh==1.8.1
pdfminer.six==20240706
pillow==11.0.0
pip==24.2
protobuf==5.29.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.14
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.8.1
safehttpx==0.1.1
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
tomlkit==0.12.0
torch==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
triton==3.1.0
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==12.0
wheel==0.44.0
